- Block delete post if post = motd post id; - Middleware FillMotd + motd.json + motd in config; - Settings controller (+ view) to index and update motd json; - Increase post title max characters

+++ api/app/Http/Controllers/Admin/PostController.php
@@ -58,8 +58,15 @@ class PostController extends Controller
 
     public function destroy(Post $post){
         $this->authorize('destroy', $post);
-        $post->delete();
-        return response()->json([], 200);
+        if($post->id != config('dofus.motd.postid'))
+        {
+            $post->delete();
+            return response()->json([], 200);
+        }
+       else
+       {
+           return response()->json(['motd' => ['0' => 'Can\'t delete this post (MOTD)']], 422);
+       }
     }
 
     public function edit(Post $post)
